Auth Routes and Auth Controller, with working Sign-in and login is almost done
