+++ server/package.json
@@ -15,6 +15,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^3.0.3",
     "concurrently": "^4.1.0",
     "express": "^4.16.4",
     "express-session": "^1.15.6",
