+++ server/server.js
@@ -58,72 +58,6 @@ app.get('/cutomerWithArtistsPictures', (req, res) => {
     console.log(customerArtist);
   })
 })
-app.post('/customerOrArtist', (req, res) => {
-
-  if (req.body.type === "customer") {
-    db.Customer.create(req.body).then(function (dbData) {
-      var userObj = {
-        _id: dbData._id,
-        firstName: dbData.firstName,
-        lastName: dbData.lastName,
-        email: dbData.email,
-        phone: dbData.phone,
-        type: dbData.type
-      }
-      req.session.customer = userObj;
-      req.session.customer.loggedIn = true;
-      res.json(dbData);
-    })
-  } else {
-
-    var customerObj = {
-      type: req.body.type,
-      firstName: req.body.firstName,
-      lastName: req.body.lastName,
-      phone: req.body.phone,
-      email: req.body.email,
-      password: req.body.password
-    };
-    console.log(customerObj);
-    db.Customer.create(customerObj).then(function (customerData) {
-      var artistData = {
-        specialization: req.body.artistData.specialization,
-        pricing: req.body.artistData.pricing,
-        location: req.body.artistData.location,
-        address: req.body.artistData.street,
-        city: req.body.artistData.city,
-        state: req.body.artistData.state,
-        zip: req.body.artistData.zip,
-        Customer_Id: customerData._id
-      }
-      db.Artist.create(artistData).then(function (artistData) {
-        console.log(artistData);
-        var userObj = {
-          _id: customerData._id,
-          firstName: customerData.firstName,
-          lastName: customerData.lastName,
-          email: customerData.email,
-          phone: customerData.phone,
-          type: customerData.type,
-          artistData: {
-            artistId: artistData._id,
-            specialization: artistData.specialization,
-            pricing: artistData.pricing,
-            location: artistData.location,
-            street: artistData.street,
-            city: artistData.city,
-            state: artistData.state,
-            zip: artistData.zip
-          }
-        };
-        req.session.customer = userObj;
-        req.session.customer.loggedIn = true;
-        res.json(userObj);
-      })
-    })
-  };
-})
-
 //should this be in the API routes/controller files (or does it have to be here)
 //also shouldn't all of the auth routes, be in the authRoutes file/controller
 //or because it uses Express Sessions it has to be in server file?
